// frontend/src/types/dashboard/contacts/driver.ts
export type Driver = {
  id: number;
  name: string;
  email: string;
  contactNo: string;
  address: string;
  active: boolean;
  avatar: string;
  dateOfBirth: string;
  maritalStatus: string;
  gender: string;
  bloodGroup: string;
  emergencyNumber: string;
  licenseExpDate: string;
  licenseIssueDate: string;
  licenseNumber: string;
  referenceBy: string;
  licensePhoto: string;
};